requirements.txt
numpy
matplotlib
tensorflow
pillow